Rename file to reflect its implementation date
These files are related more to the test suite than the schema version,
although obviously are dependent on the schema.

Also note this file was accidentally included in a commit of other test changes
and so the history is incorrect.  At that point it wasn't decided that this
implementation of upgrades would go ahead.