Simplify i18n handling in Javascript
Detect language on server side and pass that to templating.  Load language file
based on that variable which is a simple lookup table of translation strings,
and define a couple of basic functions to handle on-the-fly translations as
well as translation of static DOM elements at load.

Do not use JSON, which does not support comments; do not load language file in
asynchronously because you need it right away.